requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.6.3
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.13.3
blinker==1.8.2
cachetools==4.2.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
circuitbreaker==1.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
crispy-bootstrap5==0.7
cryptography==41.0.6
cssselect2==0.7.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
Django==5.0.6
django-bootstrap5==24.2
django-crispy-forms==2.0
django-extensions==3.2.3
django-filer==3.1.3
django-filter==23.2
django-polymorphic==3.1.0
django-taggit==4.0.0
djangorestframework==3.14.0
easy-thumbnails==2.8.5
et-xmlfile==1.1.0
fastapi==0.110.1
ffmpy==0.3.2
filelock==3.13.3
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
google-api-core==2.19.0
google-api-python-client==2.133.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.1
gradio==4.25.0
gradio_client==0.15.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.4
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==6.8.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langsmith==0.1.67
lxml==5.2.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
moviepy==2.1.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
oci==2.127.0
openpyxl==3.1.3
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pillow==10.3.0
proglog==0.1.12
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.7.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.9
python-telegram-bot==13.15
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.34.0
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
ruff==0.3.5
scapy==2.6.1
semantic-version==2.10.0
setuptools==80.4.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.6
SQLAlchemy==2.0.30
sqlparse==0.4.4
starlette==0.37.2
svglib==1.5.1
tenacity==8.3.0
tinycss2==1.3.0
tk==0.1.0
tomlkit==0.12.0
toolz==0.12.1
tornado==6.1
tqdm==4.66.2
typer==0.12.1
typing_extensions==4.7.1
tzdata==2024.1
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.29.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.15.0
